backend/requirements.txt
django==5.2.3
django-ninja==1.4.3
pydantic==2.11.7
uvicorn==0.34.3
redis==6.2.0
PyJWT==2.10.1
